@@ -1,10 +1,10 @@
 import test from 'ava'
 
-import queue = require('..')
+import queue from '..'
 
 // Tests
 
-test('should return bull instance with namespace', (t) => {
+test('should return bull instance with namespace', t => {
   const bull: any = {}
 
   const q = queue({ queue: bull })
@@ -13,7 +13,7 @@ test('should return bull instance with namespace', (t) => {
   t.is(q.namespace, 'great')
 })
 
-test('should get queue namespace from options', (t) => {
+test('should get queue namespace from options', t => {
   const options = { namespace: 'greater' }
 
   const q = queue(options)
@@ -21,7 +21,7 @@ test('should get queue namespace from options', (t) => {
   t.is(q.namespace, 'greater')
 })
 
-test('should create queue', (t) => {
+test('should create queue', t => {
   const options = { namespace: 'greater' }
 
   const q = queue(options)
@@ -29,7 +29,7 @@ test('should create queue', (t) => {
   t.truthy(q.queue)
 })
 
-test('should create queue with redis url', (t) => {
+test('should create queue with redis url', t => {
   const options = { namespace: 'greater', redis: 'redis://localhost:6378' }
 
   const q = queue(options)
@@ -37,8 +37,11 @@ test('should create queue with redis url', (t) => {
   t.truthy(q.queue)
 })
 
-test('should create queue with redis options', (t) => {
-  const options = { namespace: 'greater', redis: { host: 'localhost', port: 6378 } }
+test('should create queue with redis options', t => {
+  const options = {
+    namespace: 'greater',
+    redis: { host: 'localhost', port: 6378 }
+  }
 
   const q = queue(options)
 
